+++ src/components/RELAY/RelayChannel.h
@@ -5,18 +5,20 @@
 #ifndef RELAYCHANNEL_h
 #define RELAYCHANNEL_h
 
+#include <stdint.h>
+
 struct RelayChannel
 {
-    int relay_pin;
-    int channel_state;
+    uint8_t relay_pin;
+    uint8_t channel_state;
 
     RelayChannel();
     ~RelayChannel();
-    RelayChannel(int relay_pin, int channel_state);
+    RelayChannel(uint8_t relay_pin, uint8_t channel_state);
 
-    int get_channel_state();
-    int get_channel_pin();
-    void set_channel_state(int state);
+    uint8_t get_channel_state();
+    uint8_t get_channel_pin();
+    void set_channel_state(uint8_t state);
 };
 
 #endif /* RELAYCHANNEL_h */
